* apt-pkg/pkgcache.h:   - switch {,Install-}Size to unsigned long long * apt-pkg/depcache.cc:   - deal with long long, not with int to remove 2GB Limit (LP: #250909) (cherry picked from commit e2c66de5c5e63d8400efb0522c31fbe1ec225f93)

+++ apt/apt-pkg/acquire-item.h
@@ -49,7 +49,10 @@ class pkgAcquire::Item
     * Do not use anonyomus enums, as this breaks SWIG in some cases */
    enum StatusFlags {StatIdle, StatFetching, StatDone, StatError} Status;
    string ErrorText;
-   unsigned long FileSize;
+
+   /** \brief The size of the object to fetch. */
+   unsigned long long FileSize;
+
    unsigned long PartialSize;   
    const char *Mode;
    unsigned long ID;
